billing_kill_switch/requirements.txt
functions-framework==3.*
google-cloud-billing==1.*